Move window provider to shared module
