+++ src/app/projects/projects-list/projects-list.component.spec.ts
@@ -5,7 +5,7 @@ import { ProjectsService } from '../projects.service';
 import { FiltersComponent } from '../components/filters/filters.component';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { AnalyticsService } from 'src/app/services/analytics.service';
-import { WINDOW_PROVIDERS } from 'src/app/services/window.provider';
+import { WINDOW_PROVIDERS } from 'src/app/shared/services/window.provider';
 
 describe('ProjectsListComponent', () => {
   let component: ProjectsListComponent;
