+++ src/app/services/analytics.service.ts
@@ -1,14 +1,14 @@
 import { Inject, Injectable } from '@angular/core';
 import { Project } from '../projects/project';
-import { WINDOW } from './window.provider';
+import { WINDOW } from '../shared/services/window.provider';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class AnalyticsService {
   window: Window;
 
-  constructor(@Inject(WINDOW) window: Window) { 
+  constructor(@Inject(WINDOW) window: Window) {
     this.window = window;
   }
 
@@ -16,12 +16,16 @@ export class AnalyticsService {
     const LIMIT = 5;
 
     this.window['dataLayer'].push({
-      event: "search_projects",
+      event: 'search_projects',
       search_term: $search,
-      search_projects_term_and_uuids:
-        `term: ${$search}, results: ${projects.slice(0, LIMIT).map(dataPoint => dataPoint.uuid).join(",")}`,
-      search_projects_term_and_titles:
-        `term: ${$search}, results: ${projects.slice(0, LIMIT).map(dataPoint => dataPoint.title).join(",")}`,
+      search_projects_term_and_uuids: `term: ${$search}, results: ${projects
+        .slice(0, LIMIT)
+        .map((dataPoint) => dataPoint.uuid)
+        .join(',')}`,
+      search_projects_term_and_titles: `term: ${$search}, results: ${projects
+        .slice(0, LIMIT)
+        .map((dataPoint) => dataPoint.title)
+        .join(',')}`,
     });
   }
 }
