+++ src/app/shared/services/window.provider.ts
